+++ changelog.md
@@ -18,6 +18,11 @@
 
 
 
+*	2020.6.21
+
+	+	lilyDocument.ts: makeScoreV3 implemented.
+
+
 *	2020.6.20
 
 	+	lilyDocument.ts: containsRepeat added.